api/pyproject.toml
[project]
name = "marrow-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "python-multipart>=0.0.9",
    "authlib>=1.3",
    "PyJWT>=2.8",
    "httpx>=0.28",
    "itsdangerous>=2.1",
]

[project.scripts]
marrow = "marrow.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "ruff>=0.9",
]

[tool.setuptools.packages.find]
include = ["marrow*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
